rename MatchResolver to MatchService for consistency
